Major updates, including some portions of !me command
Some bugfixes, bot now tracks users in channel. A portion of the !me
command was added, allowing users to configure how the bot responds to
some commands via a web interface. More to come later.

Currently an issue where bot is lagging on occasion for up to 20-30
seconds. Not sure why.